Added proper validations and ResponseStatus codes.

+++ src/main/java/com/mohit/UserProvisioning/Service/RoleService.java
@@ -1,8 +1,11 @@
 package com.mohit.UserProvisioning.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.mohit.UserProvisioning.Entity.Role;
@@ -14,31 +17,52 @@ public class RoleService {
 	@Autowired
 	RoleRepository roleRepository;
 
-	public Role createRole(Role role) {
+	public ResponseEntity<String> createRole(Role role) {
 
-		return roleRepository.save(role);
+		roleRepository.save(role);
+		return new ResponseEntity<>("Role Created Successfully", HttpStatus.CREATED);
 	}
 
-	public List<Role> getRoles() {
+	public ResponseEntity<List<Role>> getRoles() {
 
-		return roleRepository.findAll();
+		List<Role> list=roleRepository.findAll();
+		return new ResponseEntity<>(list, HttpStatus.OK);
 	}
 
-	public Role getRoleById(Long id) {
-		return roleRepository.findById(id).orElse(null);
+	public ResponseEntity<?> getRoleById(Long id) {
+		Optional<Role> userOptional = roleRepository.findById(id);
+		if (userOptional.isPresent())
+			return new ResponseEntity<>(userOptional, HttpStatus.OK);
+		else {
+			return new ResponseEntity<>("Role not found with id:" + id, HttpStatus.NOT_FOUND);
+		}
 	}
 
-	public Role updateRole(Long id, Role role) {
-		Role existingRole = getRoleById(id);
-		existingRole.setName(role.getName());
-		existingRole.setDescription(role.getDescription());
-		return roleRepository.save(existingRole);
+	public ResponseEntity<?> updateRole(Long id, Role role) {
+		Optional<Role> roleOptional = roleRepository.findById(id);
+		if (roleOptional.isPresent()) {
+			Role existingRole = roleOptional.get();
+			existingRole.setName(role.getName());
+			existingRole.setDescription(role.getDescription());
+			Role savedRole = roleRepository.save(existingRole);
+			return new ResponseEntity<>(savedRole, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>("Role not found with id:" + id, HttpStatus.NOT_FOUND);
+		}
+
 
 	}
 
-	public void deleteRole(Long id) {
-		roleRepository.deleteById(id);
+	public ResponseEntity<?> deleteRole(Long id) {
+		Optional<Role> userOptional = roleRepository.findById(id);
+		if (userOptional.isPresent()) {
+			roleRepository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		}
 
+		else {
+			return new ResponseEntity<>("Role not found with id:" + id, HttpStatus.NOT_FOUND);
+		}
 	}
 
 }
